requirements.txt
certifi>=2021.5.30
importlib_metadata>=4.8.1
qiskit-aer>=0.10.3
qiskit-terra>=0.23.0
matplotlib>=3.3
